[unify-go-to-next-location] Expose lsp-{next,previous}-location to traverse any grep-like buffer
These commands work for any grep-like buffer - the user just needs to pass
`%val{bufname}` as first argument for example `*goto*`, or `*grep*` [1].

I find this very convenient because it makes it trivial to define a command
that allows to go to the next/previous location in any kind of grep buffer.
This works for all lsp buffers with filetype `lsp-goto` as well as for `grep`,
`lint` and `make` buffers:

	declare-option -hidden str last_used_grep_buffer
	hook -group my global WinDisplay \
		\*(?:diagnostics|goto|find|grep|implementations|lint-output|make|references|symbols)\* %{
		set-option global last_used_grep_buffer %val{bufname}
	}
	map global normal <c-n> %{: lsp-next-location %opt{last_used_grep_buffer}<ret>} -docstring 'lsp next'
	map global normal <c-p> %{: lsp-previous-location %opt{last_used_grep_buffer}<ret>} -docstring 'lsp previous'

This was originally posted as
https://discuss.kakoune.com/t/single-command-for-grep-next-match-in-similar-buffers-lsp-make-find/1215
I wonder if something like this should be the default in future.

[1]: Or `%val{buffile}` which is equivalent to `%val{bufname}` for `*grep*` and friends.